requirements.txt
streamlit 
pytesseract 
pillow
pyperclip
google-cloud-vision
transformers
logging
# easyocr